refactor(web): move the indexing selection helpers to types and utils
The model selection types now sit with the other indexing types, and the
selection helpers move to a new utils module. Every sibling library
already keeps its helpers in utils.ts, so indexing was the exception.

This only moves code. Behaviour is unchanged.
